Update your table definition to include the park description. Update your insert statements to use $dbc->prepare() and bound variables. You may still use $dbc->exec() to create the table.
Update the query(s) in national_parks.php to use prepared statements, in particular for the limit and offset.

Add a form that allows national parks to be added to your database. Use prepared statements for all queries that contain dynamic data.

Validate all form inputs to ensure that database entries are not empty.